Added in baseline validation layer configuration
The program now checks for validation layer
support (if requested via preprocessor checks)
and sets up the validation layer plugin to
use the debugCallback(...) function defined
on line 73.

This is done by using the vkCreateDebugUtilsMessengerEXT
function, which must be loaded through the
vkGetInstanceProcAddr function since vulkan extension
functions aren't exported in the global namespace.
This vkCreateDebugUtilsMessengerEXT function takes
in a VkDebugUtilsMessengerCreateInfoEXT struct
which configures the debug levels and the callback
function, and allocates
a messenger object for the current instance.

This same structure is also passed into the
pNext attribute in the VkInstaceCreateInfo struct
so that the same callback can be used upon
instance creation and destruction.

There are a few helper functions that wrap
all of the tedious and ugly struct population
and dynamic function loading so that the code
is a bit readable, but it's still a bit of a
mess.

The next tutorial will have the program select
physical devices.