Select a device based on it's queue properties
device_is_suitable() will wrap around the
vkGetPhysicalDeviceQueueFamilyProperties() function
in order to query a device's queue support.
If a VkPhysicalDevice's VK_QUEUE_GRAPHICS_BIT
is set, we then set the index in a queue_family_indices
struct which will keep track of where the queue resides
in the queueFamilies vector in order to be used by the
program.