Added in window surface and present logical device
The window surface is created using the
glfwCreateWindowSurface()
which comes from the glfw vulkan extensions.
This allows us to rely on glfw to actually do
whatever platform specific functions in order
to link the vulkan infrastructure into the glfw
window.

The present logical device links to the specific
queue family that supports presenting to the surface
that we created using glfw. This then gives us
a queue family that will display to the window surface.