Added in shaders and default pipeline structs
I've added the shader src directory in the
Makefile as $(SPV) and $(DSPV) respectively.
I originally wanted to just build them separately
and symlink them, however ifstream doesn't follow
symlinks, and making it follow symlinks requires
using some ugly iostream hacks.

There's a lot of boilerplate struct filling out
here, and they currently aren't used (this is
inline with a major step of the tutorial).

I've commented in the details of each config struct,
broadly speaking there are two parts to the
pipeline, programmable and non-programmable.
The non programmable parts are non programmable
due (presumably) to the reliance on hardware acceleration
for these steps IE: the rasterizer

A future project would be to modularize the
create_graphics_pipeline function
in the future, which could be done with late
binding or something.